s2l corner_grow

+++ generators/g_corner_grow.py
@@ -36,7 +36,7 @@ class g_corner_grow():
         # check if S2L is activated
         if self.channel >= 0:
             current_volume = float(str(self.sound_values.buf[self.channel*8:self.channel*8+8],'utf-8'))
-            self.waiting = 50 - np.clip(current_volume, 0, 1) * 50
+            self.waiting = 50 - int(np.clip(current_volume, 0, 1) * 50)
 
         if self.size > 4.0:
             # switch into waiting state
